Hack patch to compile with latest libvirt in chroot-fc23
This change makes the patch an effective no-op since the define it needs
isn't present (per the comment, I presume another solution was found). Doing
This seems better than trying to find a way to not apply the patch based
on the build distro.

I didn't attempt to fully regenerate the patch from a git repo but that
shouldn't matter since it won't ever be getting merged to mainline. It still
applies OK as a frankenpatch.